requirements.txt
python-dotenv
ibm-generative-ai
ibm-generative-ai[langchain]
pypdf
InstructorEmbedding
transformers[torch]
sentence-transformers
Flask 
flask-restful 
flask_httpauth
cachetools
unstructured
psutil

faiss-cpu
evaluate
flask_cors
rogue
ibm-watson
rouge_score

gunicorn